backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2.7
numpy>=1.26
pandas>=2.0
scikit-learn>=1.3
joblib>=1.3
matplotlib>=3.8
python-multipart>=0.0.6
groq>=0.4.0
python-dotenv>=1.0.0
